Use a relative link to module order in manual
Use a relative link instead of an absolute link to the user manual in the FAQ about changing module order. Also stop hard-coding the version in the manual link and allow a redirect to the latest stable version instead.

It looks like this is more consistent with how other links to the manual are set up in this doc.